requirements.txt
certifi==2020.4.5.1
Django==2.2
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-taggit==1.2.0
importlib-metadata==1.5.0
jsonfield==3.1.0
Markdown==3.2.1
mysqlclient==1.3.14
path==13.2.0
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
swapper==1.1.2.post1
wincertstore==0.2
zipp==2.2.0
